+++ sites/demo/src/backend/demo/settings.py
@@ -198,6 +198,37 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     SOCIAL_AUTH_EDX_OAUTH2_ENDPOINT = values.Value()
     SOCIAL_AUTH_POSTGRES_JSONFIELD = False
 
+    SOCIAL_AUTH_PIPELINE = (
+        # Get the information we can about the user and return it in a simple
+        # format to create the user instance later. In some cases the details are
+        # already part of the auth response from the provider, but sometimes this
+        # could hit a provider API.
+        "social_core.pipeline.social_auth.social_details",
+        # Get the social uid from whichever service we're authing thru. The uid is
+        # the unique identifier of the given user in the provider.
+        "social_core.pipeline.social_auth.social_uid",
+        # Verifies that the current auth process is valid within the current
+        # project, this is where emails and domains whitelists are applied (if
+        # defined).
+        "social_core.pipeline.social_auth.auth_allowed",
+        # Checks if the current social-account is already associated in the site.
+        "social_core.pipeline.social_auth.social_user",
+        # Make up a username for this person, appends a random string at the end if
+        # there's any collision.
+        "social_core.pipeline.user.get_username",
+        # Create a user account if we haven't found one yet.
+        "social_core.pipeline.user.create_user",
+        # User exists, update it to superuser if needed
+        "demo.social.pipeline.user.set_super_user",
+        # Create the record that associates the social account with the user.
+        "social_core.pipeline.social_auth.associate_user",
+        # Populate the extra_data field in the social record with the values
+        # specified by settings (and the default ones like access_token, etc).
+        "social_core.pipeline.social_auth.load_extra_data",
+        # Update the user record with any changed info from the auth service.
+        "social_core.pipeline.user.user_details",
+    )
+
     # LMS
     LMS_BACKENDS = [
         {
@@ -233,8 +264,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         "family_name": "last_name",
         "locale": "language",
         "user_id": "user_id",
-        "is_staff": "is_staff",
-        "is_superuser": "is_superuser",
+        "administrator": "is_staff",
     }
 
     # Templates